modules/kafka-python-twitter-stream/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
idna==3.3
    # via requests
kafka-python==2.0.2
    # via -r requirements.in
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   tweepy
requests==2.28.0
    # via
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
tweepy==4.10.0
    # via -r requirements.in
urllib3==1.26.9
    # via requests
